fix(Controller, Home): If one object, transform it in array and reboot the keys
When the Space Manager manages 1 space, the 'apps/workspace/spaces' route return one array (php side) but javascript interprets it like an object. But, if this space is the key '9', it will not print. So, I reboot theses keys with the array_values method in 'findAll()' method.

SO: https://stackoverflow.com/questions/3401850/after-array-filter-how-can-i-reset-the-keys-to-go-in-numerical-order-starting/43657546

Resolve: https://github.com/arawa/workspace/issues/377