refactore(Controller, Service, Vue): Delete an user from a workspace or (sub-)groups
With the get method from WorkspaceService, it build an object with groupfolderService. So, in the removeUser method, I send the space object and not the spaceId. But, Javascript side, I create a deep copy with JSON.parse(JSON.stringify(object)). Please, look my source

source: https://www.samanthaming.com/tidbits/70-3-ways-to-clone-objects/#shallow-clone-vs-deep-clone