fix(vue): add ref to NcActions and this.$refs.ncAction.opened=false
This PR is intended to fix a minor bug of the UI: wWhen trying to rename a user group and there is an error message
the modal window continue to be present on the page. This is fixed with adding ref to corresponding component and then call 'this.$refs.ncAction.opened=false'.