hotfix(php): Ignore guest virtual group
When we run the update of Workspace to 3.0.X, the migration is stopped
and the program doesn't change the display name of Workspace's groups.

Why ? Because, when we use the "search" function from "IGroupManager",
by default, in addition to groups beginning with "SPACE-GE" or
"SPACE-U", we get the guest virtual group.

So, the solution is to use the "array_filter" to ignore the
"OCA\\Guests\\GroupBackend" backendname.

Signed-off-by: Baptiste Fotia <fotia.baptiste@hotmail.com>