feat(php): Add the ability to add groups in workspaces
We can add connected groups (or added groups) to workspaces. These groups can be LDAP or local, and when we add a connected group, we also add its users. In this feature, we return a slug of a gid. This slug handles special characters, making it more suitable for rendering in the browser's URL.