refactor(middleware): Every Middleware has one responsibility to manage a response
Now, each Middleware is responsible for one use case.

For example, when creating a workspace with special character, the WorkspaceNameSpecialCharException will be thrown, and the SpacenameForbiddenCharactersMiddleware will be generate a response depending on whether the exception comes from an OCSController.

Tomorrow, if a specific exception needs to be thrown, we will create that exception to be thrown from a controller or
service, with a corresponding Middleware that executes an afterException function and returns a suitable response with the error message based on its origin.

OP#4353
OP#4354
OP#4355
OP#4356
OP#4357
OP#4358